단어세기
텍스트가 입력으로 주어질 때, 단어의 개수를 세는 프로그램을 작성한다.
Write a program that counts words when text is given as input.

아래 내용을 가진 텍스트파일을 미리 만들어 두고, 프로그램을 실행하면 파일 내용을 읽어들인다(출처: Wikipedia).
Create a text file with the following contents in advance, and execute the program to read the contents of the file (source: Wikipedia).

As the country became embroiled in a domestic crisis, the first government was dislodged and succeeded by several different administrations. Bolikango served as Deputy Prime Minister in one of the new governments before a partial state of stability was reestablished in 1961. He mediated between warring factions in the Congo and briefly served once again as Deputy Prime Minister in 1962 before returning to the parliamentary opposition. After Joseph-Desire Mobutu took power in 1965, Bolikango became a minister in his government. Mobutu soon dismissed him but appointed him to the political bureau of the Mouvement Populaire de la Revolution. Bolikango left the bureau in 1970. He left Parliament in 1975 and died seven years later. His grandson created the Jean Bolikango Foundation in his memory to promote social progress. The President of the Congo posthumously awarded Bolikango a medal in 2005 for his long career in public service.

구분자(Separator)는 마침표 '.', 쉼표 ',', 공백 ' ' 이다.
가장 많이 나온 순서대로 단어 10개와 그 단어의 빈도를 출력한다.
빈도가 같은 단어들 사이의 순서는 무시한다.

Separators are period'.', comma',', and space ''.
It prints 10 words and the frequency of those words in the order in which they appear most.
Ignore the order between words of the same frequency.

in 12
the 10
Bolikango 5
a 4
of 4
and 3
to 3
his 3
became 2
government 2